Phase 3: Consolidate shim layers into canonical modules
The V2 refactor left duplicate one-line re-export layers (simulation/, rl/,
envs/runtime_state.py, envs/step_info.py, root factory.py, config/defaults.py,
heuristics/first_fit.py) shadowing the canonical runtime/, features/, api/
modules — plus two pieces of real code stranded inside shim packages. Per plan
Phase 3, collapse everything onto the canonical paths.

WHAT changed:
- Move the stranded compiled kernel simulation/_request_analysis_kernels.pyx to
  runtime/_request_analysis_kernels.pyx; update the Extension entry in setup.py
  and the try-import in runtime/request_analysis.py.
- Move the stranded real RewardFunction implementation rl/reward_function.py to
  features/reward_function.py (its siblings rl/action_mask.py and
  rl/observation.py were star re-exports of features/); export it from
  features/__init__.py.
- Delete the shim modules: simulation/ (entire dir), rl/ (entire dir),
  envs/runtime_state.py, envs/step_info.py, factory.py (root),
  config/defaults.py, heuristics/first_fit.py. optical/first_fit_example.py is
  kept: it holds the real run_episode (verified, not a shim).
- Repoint consumers to canonical modules: bench/benchmarking.py and
  runtime/simulator.py (features.reward_function), api/factory.py and
  config/__init__.py (optical_networking_gym.defaults), optical/
  first_fit_example.py (api.factory, heuristics.masked_heuristics),
  tests/observation/test_request_analysis_hot_path_helpers.py
  (runtime.request_analysis), and the RewardFunction entry in the package
  __init__ export table.
- Trim envs/__init__.py to export only OpticalEnv (its other entries pointed at
  the deleted shims; no consumers imported them from envs).

WHY: one canonical module per component; no more double import paths for the
same class, and the compiled kernel now lives next to the code that uses it.

Verification gate:
- pytest: 191 passed, 0 skipped
- python -c "import optical_networking_gym as g; [getattr(g, n) for n in g.__all__]"
  succeeds (83 exports)
- compiled kernel verified to load from optical_networking_gym.runtime
  (_request_analysis_kernels .so, not the pure-Python fallback)
