+++ src/optical_networking_gym/bench/benchmarking.py
@@ -22,7 +22,7 @@ from optical_networking_gym.network.topology import TopologyModel
 from optical_networking_gym.optical.qot_engine import QoTEngine
 from optical_networking_gym.features.action_mask import ActionMask
 from optical_networking_gym.features.observation import Observation
-from optical_networking_gym.rl.reward_function import RewardFunction
+from optical_networking_gym.features.reward_function import RewardFunction
 from optical_networking_gym.runtime.request_analysis import RequestAnalysisEngine
 from optical_networking_gym.config.scenario import ScenarioConfig
 from optical_networking_gym.stats.statistics import Statistics
