--- src/optical_networking_gym/config/defaults.py
@@ -1,3 +1,0 @@
-from optical_networking_gym.defaults import MODULATION_CATALOG, get_modulations, resolve_topology, set_topology_dir
-
-__all__ = ["MODULATION_CATALOG", "get_modulations", "resolve_topology", "set_topology_dir"]
