+++ src/optical_networking_gym/envs/__init__.py
@@ -3,10 +3,7 @@ from __future__ import annotations
 from importlib import import_module
 
 _EXPORTS = {
-    "ActiveService": (".runtime_state", "ActiveService"),
     "OpticalEnv": (".optical_env", "OpticalEnv"),
-    "RuntimeState": (".runtime_state", "RuntimeState"),
-    "StepInfo": (".step_info", "StepInfo"),
 }
 
 __all__ = list(_EXPORTS)
