--- src/optical_networking_gym/envs/runtime_state.py
@@ -1,1 +1,0 @@
-from optical_networking_gym.runtime.runtime_state import *  # noqa: F403
