--- src/optical_networking_gym/envs/step_info.py
@@ -1,1 +1,0 @@
-from optical_networking_gym.runtime.step_info import *  # noqa: F403
