--- src/optical_networking_gym/factory.py
@@ -1,3 +1,0 @@
-from optical_networking_gym.api.factory import make_env
-
-__all__ = ["make_env"]
