--- src/optical_networking_gym/heuristics/first_fit.py
@@ -1,3 +1,0 @@
-from optical_networking_gym.heuristics.masked_heuristics import select_first_fit_action
-
-__all__ = ["select_first_fit_action"]
