--- src/optical_networking_gym/rl/action_mask.py
@@ -1,1 +1,0 @@
-from optical_networking_gym.features.action_mask import *  # noqa: F403
