--- src/optical_networking_gym/rl/observation.py
@@ -1,1 +1,0 @@
-from optical_networking_gym.features.observation import *  # noqa: F403
