--- src/optical_networking_gym/simulation/action_codec.py
@@ -1,1 +1,0 @@
-from optical_networking_gym.runtime.action_codec import *  # noqa: F403
