--- src/optical_networking_gym/simulation/request_analysis.py
@@ -1,15 +1,0 @@
-from optical_networking_gym.runtime.request_analysis import (  # noqa: F401
-    RequestAnalysis,
-    RequestAnalysisEngine,
-    _analyze_free_mask,
-    _fragmentation_damage_by_candidates,
-    _summary_after_allocation,
-)
-
-__all__ = [
-    "RequestAnalysis",
-    "RequestAnalysisEngine",
-    "_analyze_free_mask",
-    "_fragmentation_damage_by_candidates",
-    "_summary_after_allocation",
-]
