--- src/optical_networking_gym/simulation/scenario.py
@@ -1,3 +1,0 @@
-from optical_networking_gym.config.scenario import ScenarioConfig
-
-__all__ = ["ScenarioConfig"]
