--- src/optical_networking_gym/simulation/simulator.py
@@ -1,1 +1,0 @@
-from optical_networking_gym.runtime.simulator import *  # noqa: F403
