--- src/optical_networking_gym/simulation/traffic_model.py
@@ -1,1 +1,0 @@
-from optical_networking_gym.runtime.traffic_model import *  # noqa: F403
